refactoring code tables and functions
