+++ constants/ProgramCodes.js
@@ -1,0 +1,6 @@
+export const ProgramCodes = {
+  CPP: 'CPP',
+  EI: 'EI',
+  OAS: 'OAS',
+  GIS: 'GIS',
+}
