+++ constants/StatusCodes.js
@@ -1,0 +1,7 @@
+export const StatusCodes = {
+  inPayment: 'inPayment',
+  benefitUpdate: 'benefitUpdate',
+  applicationReceived: 'applicationReceived',
+  inactive: 'inactive',
+  exahusted: 'exhausted',
+}
