+++ constants/SummaryTypes.js
@@ -1,0 +1,6 @@
+export const SummaryTypes = {
+  PaymentAmount: 'PaymentAmount',
+  NextReport: 'NextReport',
+  LatestStatus: 'LatestStatus',
+  NextPayment: 'NextPayment',
+}
