+++ pages/api/programData/_middleware.js
@@ -1,0 +1,41 @@
+import { TASK_GROUPS } from '../../../contents/DashboardBenefitTasksConstants'
+
+export function CreateGenericBenefitObj(
+  programCode,
+  statusCode,
+  typeCode,
+  summaries
+) {
+  const benefit = {
+    programCode: programCode,
+    statusCode: statusCode,
+    typeCode: typeCode,
+    summaries: summaries, // an array of BenefitSummaries
+  }
+  const taskGroupMatch = MapBenefitToTaskGroups(programCode, statusCode)
+  if (taskGroupMatch) {
+    benefit.taskGroups = taskGroupMatch.tasksGroups
+    benefit.taskHeadingKey = taskGroupMatch.taskHeadingKey
+  }
+  return benefit
+}
+
+//Programs
+function MapBenefitToTaskGroups(programCode, statusCode) {
+  const group = TASK_GROUPS.find(
+    (tg) => tg.programCode === programCode && tg.statusCode === statusCode
+  )
+  return group
+}
+
+export function CreateBenefitSummary(type, value, status) {
+  let benefitSummary = {
+    type: type, // will define what header text to go with it and any links if applicable
+    value: value, // a date or amount, defined by the type
+  }
+  //only add the status object if it exists
+  if (status) {
+    benefitSummary.status = status //status is additional text for display
+  }
+  return benefitSummary
+}
